requirements.txt
python-binance==1.0.16
python-kucoin==2.1.3
btcturk_api==1.8.1
huobiApi==0.0.5
matplotlib==3.5.1
mplcursors==0.5.1
numpy==1.22.1
requests==2.22.0
websocket_client==1.2.3
